Optimize minute and hour increments
Even when the minutes and hour expression were specified, CronExpression would iterate through every minute and every hour in order to hunt for a valid date/time for the given expression. This would cause a lot of iterations for some events (such as those occuring daily) and might even cause failure for $nth > 0.

Instead of iterating every minute or hour for given value, try to jump to the next suitable value which severly limits the number of iterations required.